seperate choices from models
